Move repository actions to page title for viewports width below 1280px
